refactor: split pageBlocks into definitions, responses, guards
523 lines held the block registry, 25 per-block response types, and the
guards/identity/label helpers, so finding "what blocks exist" meant
scrolling past a wall of response shapes.

Split along those three seams and keep pageBlocks.ts as a barrel (523 ->
61 lines); 31 modules import from it.

Used flat sibling filenames (pageBlocks.definitions.ts etc.) to match the
existing convention in this directory (homepageFeaturedSlots.utils.ts,
locationGridTypes.ts) and to avoid a pageBlocks/ directory making
'./pageBlocks' ambiguous.

The two derived Sets backing the guards are now exported from the
definitions module rather than recomputed, so they stay tied to the
registry they are derived from.

Verified: export surface identical (43/43 names via the TS checker,
diffed against main); response types copied verbatim with sed; tsc error
signature identical to baseline; vitest 333 pass / 3 pre-existing
failures, same as main.
